ops-prof: add disabled (slow) testcase for full set of operations.
fails somewhere around half-way on an assert in codegen for an
out_chans=384 case. it looks like the usual sort of not-handled case,
so we should probably make the error non-fatal so that the case
doesn't kill the whole run, as with the (Kb==1) assert before
(although we now have in thoery 'fixed' that, since it should have
never happened anyway -- it was due to using an sgemm-flavor op_tune
with a conv op).