comp_util: oops, replace get_diff_double() with get_rd_double()
if we're determining errors based on our relative-difference metric
(aka 'rd'), we need to use that same metric to determine which
elements to print out. so, that was certainly a bug, and it explains
why the failing cases were outputing large, nearly equal values as
example failures -- these points had large *absolute* difference, but
not large relative difference. with the correction, we see a (single)
more-reasonalbe failing point for the conv-debug failing ops-prof case
extracted from the conv-full set.