conv-full-gen5 test case: update failing results after get_diff_double()->get_rd_double() fix
now, we get a more reasonable/correct set of example failing
locations; in particular they are often pretty sparse and all for
small values, which might explain why there are no digest
failures. for digests that sum over many points, the small errors
would be washed out. and for the digest points that have only one or a
few samples, it's possible they simply don't happen to hit any failing
points, since there often seem to be only a few of them. however, some
cases seem to have a pretty large number of failure and no digest
failure, so we need to investigate more.