ops-prof: use specified-by-tag op_tune for comparisons and digests
as opposed to taking the first non-failing one. however, since we need
the results of the known-good tune to compare against other tunes, we
had to split the for-loop over the tunes and cache all the per-tune
results. see the relevant FIXME. but, this should be
more-consistent/sane overall. update the currently-failing test
results for conv-debug (proxy for conv-full).