ops-prof: update slow conv-full test wrt prior change
as expected, the cases that were previously using cudnn as the
baseline (due to the '8...' tune failing) now are
(presumably/hopefully) using ocl-def as the baseline. thus, instead of
almost every tune failing as before, only the cudnn tune fails (in
line with other tests). now, the majority (all?) failures are
vs. cudnn, with errors ranging from a bit-more-that-the-current-rd
(current rd is 2e-4, many errors are ~4e-4), to a few tests with at
least a few points that have up to ~2e-3 error (seems to be 3x3
winograd cases). however, nothing really looks wrong; the next TODO is
to set a higher toler (~4e-4) for the cudnn cases, hopefully splitting
out the winograd ones somehow (which need 2e-3 (!)).

note that, unfortunatly, the digests certainly aren't too good at
catching a small % errors that are above tolerance (although this is
expected). for now, we'll accept this. in general, we might want to
consider some kind of statistical tolerance, were we allow a few
larger errors, or many smaller ones. then, the digest and live-compare
would probably match up better, and that might be okay/better as an
error metric? hmm.