pass reduced-function-signature-op back from profile_rcg_call
and in ops-prof, add this op to the per-run wisdom. it makes sense to
show exactly what function acutally was run, but this was added in
particular so we can use the function name to adjust the rd tolerance
for cudnn. still TODO is some way to know cudnn algo is used in each
case, but note that this is dynamic, so it's a bit unclear how we'd
get that info back. that is, it's some meta-return-value from running
the function, not something we know at compile/gen time. pending a
good/clean solution, we're also pondering suitable hacks in the mean
time ... maybe something like upping the tolerance for 3x3 convs where
we know/guess winograd will be selected. another alternative is to
break out all the cudnn sub-algos, which we want to do anyway ...