tweak to boda_make_util to actually preserve obj_list comments
this might have been a bug, but we might have added more trying to fix
it :/ ... for the android obj_list, there was a line that continued to
a line with a comment, and the comment was being removed, which isn't
right. however, the following line in the final obj_list was (and had
to be?) a comment as well, so there was no observable effect. but,
keeping the comments seems okay/more-right.