profile_rtc_call: add bool arg to control outputting inputs
in addition to outputting the outputs of the called function, if the
bool is set, the inputs will also be output. previously, this was done
be un-commenting a line manually, but the line had gone stale, and
when we fixed it we made it a bit more official. however, no callers
currently set the bool to 1, and note that doing so in, say, ops-prof,
will invalidate all wisdom files. so, it's still mainly for
one-off/manual use. but the option could probably be added at the
ops-prof level and passed down. ehh, it's step in the right direction
anyway.