ops-with adding/fixing handling of odd Qualcomm OpenCL compile error
on the SD820/qualcomm OpenCL impl, compilation sometimes fails with
'CL_OUT_OF_HOST_MEM' (-2), but no log. maybe it's related to shared
memory usage? in any event, we had some handling for this for the
older cnn_op_info mode, but it got broken and/or never worked for
ops-prof. we've attempted to fix it back up at least for ops-prof mode
here.