wis-ana/wis-plot: various formatting/color tweaks
note: for now, wis-plot is skipping ops with less than 10e8 flops;
this is perhaps not a sensible default, but happens to be what we want
in these cases. the threshold is an option, but it's never set
(i.e. the default 10e8 thresh is always used). also, the titles are
currently being omitted from the output images, so maybe we need a way
to generate the latex snipped that actually includes all figures too
or the like.