add magic_corruption feature to data_stream_test_gen
and, sure enough, it finds a bug in data_stream_mxnet_brick magic
handling. not too exciting, but one that certainly needed at least a
single test that had a magic-val in the stream to find: ret.sz wasn't
set (ever) in the multi-part case, so it could never have
worked. here's hoping there's no actually-subtle bugs in the magic
handling. we could certainly try harder to hit some of the zero-size
cases and such.