add a num_to_proc to scan-data-stream-multi
we never resolved either the multi/non-multi split, the src/sink split
(i.e. do we need both push/pull as seperate things?). nor did we
really dealt with how start-stop-skip should interact with these. but
for now, we actually need this feature, and it's similar to one plan
we had for dealing with start-stop-skip anyway.