add data_stream_pipe, use it instead of src+sink in scan
also, related fixes/semantic changes/tweaks, and some left-over bug fixes from the prior folding of data_sink into data_stream.

we're running our tests in round-robin fashion, cleaning them up and
fixing them as we go. with automation, we can/should be running them
all each commit, but for now many of them use large inputs and are too
slow for that. but as the number of tests goes up, we'll need to
automate soon anyway ...