working toward passing velo config data to os-render
note: since we don't write the config data to bricks, we can only use
'live' mode when reading the original velo stream (i.e. when the velo
stream parsing is present). however, that mode is untested/broken for
os-render, of course. and, though we maybe were trying to avoid it, i
guess that means we'll need to read config from disk for os-render
somehow (either xml or some serialization of the stream config
data). sigh.