at long last, update goldens for change to cuda 8 / cudnn 6 / caffe head
upon investigation, it seems the changes are reasonably explained by
minor changes in the results of lrn, which has always been a source of
problems with checking numerical results. as usual, however, it's
difficult to investigate any other downstream-of-lrn changes in
results that might have occured. but, lacking any better ideas or the
will to do more in-depth investigation, we'll just update the goldens
for now. in particular, even invesitgating the changes to lrn is hard,
since we can't easily run the old version anymore, so that we could
compare a specific result and/or inspect any differences in the ptx.

note that there are no golden updates for the gen_ops_prof_tests,
though (which are per-operation for a some conv and sgemm cases), so
that's nice.