velodyne-gen: partial status byte output support
might be correct for 32-laser case; incomplete but maybe not too far
off minimal to keep veloview happy for 64 case. more work pending the
need for it and/or getting access to the veloview 3.5 source to see
what status bytes/types/values it actually parses. currently, no plans
to embed config data in 64 case though ...