data-stream: add angle-adj filter for adusting and normalizing angles in degrees
yeah, pretty special-purpose, i know. but could be reasonable extended
to handle radians and various other common angle
transforms/normalizations, so maybe it's not so horrible. but of
course it shows the bad side of the c++ vs. python/numpy bargain: it
needed a bunch of special C++ code *and* it's slow, compared to some
python that would be fast (vectorized), flexible, and need less
code. hmm.