somewhat scattered changes related to img_t YUV support
change existing usages of YUV-img-as-NDA to use (or also use) the new yuv_pels field in img_t.

minor changes to disp_win to support YUV images, but the actual data
copy support is still missing. however, disp_win now actually supports
calling disp_setup() multiple times, and also support dynamically
switching an input image with an image of the same size. this is to
work-around the fact that share_pels_from() doesn't seem to cleanly
extend to supporting our YUV hack, so we replace images
instead. debatably, the swapping of images is in general cleaner that
the share_pels_from() method, but care needs to be taken with
object/img_t identity/sharing and managment of image sizes.