experiments with raw->RGB data conversion
add decompanding, gamma correction, and post-scaling.

while some combinations of options seem interesting, and maybe-better
than the dynamic level filtering, it's not clear any of them really
match some particular raw,RGB example pairs we're considering. the
natural next step is to combine the dynamic filtering with these new
options, and/or to allow for dynamic/live experiments with changing
the values, which would make it much easier to scan around the space
visually. currently a new value takes a couple seconds to try; with
live updates it would be 1/10 of a sec or less.

also, more external data (i.e. camera metadata, or direct
experiments/analysis of the raw/RGB pairs we want) could help quite a
bit.