initial (untested) work toward supporting velo-64 data generation
i.e. pcdm->synthetic velo-64. need to extending existing velo-32
generation for 64 laser case. however, for now, we're going to
experiment with creating the needed syn-velo-64 config file that will
correspond to the output, and test that with os-render. then we'll
finish the syn-gen code, and test the resuling .pcap+config with
veloview.