add partial/semi-manual velo cfg writing
this allows for outputing the corrections section of the pcdm version
of a given velo's cfg file. this way, external tools such as veloview
should be able to interpret such pcdm data as if it came from the
sensor given by the generated config file. in particular, for pcdm
input data, the output corrections will have vertically-ordered lasers
and no horizontal angle (azimuth) corrections. tested by forcing
pcdm-mode to false and using the final generated+manually-spliced
config file in os-render (should work for velo-to-pts as
well). however, for now, we continue to have os-render/velo-to-pts
take the original/src config velo file and perform the pcdm-ification
of the config internally for now.