first pass at velo-64 syn data generation
not that we don't really try to set the status bytes. we set 'V', but
it probably doesn't matter; VeloView will tag it as 64 data based o
the the FB block ids. however, note that, maybe, if we had full
corrections in the stream, veloview might also use that to detect 64
data. but of course the S2-64 doesn't have that? but maybe some
S2-like status fields would also be detected? eh, it probably doesn't
matter too much.