os-render: make top-level pointcloud be optional
somewhat of a hack; the general idea is that raw pointclouds (and
maybe later regular ones) can be subblocks, so we may not need/want a
top-level velo-style pointcloud. but, for now, it's still sort-of
half-and-half -- there can only be only velo pointcloud, and it must
be top-level, but it is optional. on the other hand, raw point clouds
must be not-top-level.