add tools and new build config for boda deployment experiments
the new get-deps script uses ldd and "dpkg -S" to find packages and
not-in-any-package files (libs) that an exe depends on. this still
doesn't really answer the question of what the correct debian-package
deps of an exe are, since it includes transative deps, and doesn't
deal with meta-packages that might be the 'right' way to satisfy
some/multiple deps. however, it's maybe better than nothing. often, if
you use "dpkg -l" on the entire list, only a few packages might be
missing, and it's not so bad to manually install of them (or manually
determine what higher-level package(s) to install instead. but, that
said, maybe we should work more from the list of libs we like against?
or just manually maintain a list? hmm.

the obj_list.disp config has most deps not need for display disabled
to reduce dependencies. notably, caffe_pb/protoc was not disabled,
since there's a missing negative-dep for the caffe_pb-is-disabled case
currently (for generating tests). however, keeping caffe_pb and protoc
is mostly a build-time issue, and doesn't much impact deployment (just
adds a single lib for protobuf).

also, we *really* need to think about addressing the multi-config
duplication issue now.