re-enable opencl for testing
note that, along with the new cuda, we have new ndivia drivers/opencl
(nvidia-opencl-icd-387).  currently, it seems like the tests are back
to passing without error (without or without opencl), but it's not
clear if this is consistent.

note that, additionally, we saw some new failures that seem to occur
at shutdown time, inside the dtors for the caffe net (with opencl
disabled). these maybe went away after rebuilding caffe (and making
sure to use cudnn 7 instead of some random cudnn 6 from a driveworks
install), but maybe they are also just random/inconsistent.