data-stream-pcap: partial/incomplete support for IP fragmentation reassembly.
yeah, maybe should use a library for that, huh? a bit too complex as
written, and doesn't really handle enough cases to be useful, but not
too far off. with some cleanup, could probably get over the line and
handle multiple in-flight fragmented packets, or at least properly
handle non-fragmented packets interspersed with a single stream of
fragmented ones ...